# server/swagger_server/swagger/swagger.yaml
openapi: 3.0.3
info:
  title: FMD
  description: A website for movies' and TV shows' reviews
  termsOfService: http://swagger.io/terms/
  contact:
    email: om.dehlantube@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 3.0.0
servers:
- url: http://localhost:8080/VEDANTANEOGI_1/FMD_API3.0/3.0.0
...
paths:
...
  /user/{user_id}:
    get:
      tags:
      - User
      summary: Get User object(user details) from user id
      operationId: user_user_id_get
      parameters:
      - name: user_id
        in: path
        description: UserId
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          example: 1
      responses:
        "200":
          description: User details returned successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/user'
        "404":
          description: User not found
        "500":
          description: Internal Server Error
      x-openapi-router-controller: swagger_server.controllers.user_controller
  /user/signin:
    post:
      tags:
      - User
      summary: Sign in user with credentials
      operationId: user_signin_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/user_signin_body'
      responses:
        "200":
          description: User signed in successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/user'
        "400":
          description: Bad request
        "401":
          description: Invalid credentials
        "500":
          description: Internal server error
      x-openapi-router-controller: swagger_server.controllers.user_controller
  /user/signup:
    post:
      tags:
      - User
      summary: Sign Up
      operationId: user_signup_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/user_signup_body'
      responses:
        "200":
          description: User account created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/user'
        "400":
          description: Bad request
        "409":
          description: User already exists
        "500":
          description: Internal server error
      x-openapi-router-controller: swagger_server.controllers.user_controller
...
  /watchlist/{user_id}:
    get:
      tags:
      - User
      summary: Get all titles in user watchlist
      operationId: watchlist_user_id_get
      parameters:
      - name: user_id
        in: path
        description: ID of the user
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          example: 42
      responses:
        "200":
          description: List of titles in user watchlist
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/title'
                x-content-type: application/json
        "400":
          description: Bad request. Invalid user ID
      x-openapi-router-controller: swagger_server.controllers.user_controller
  /watchlist/movie/{user_id}:
    put:
      tags:
      - User
      summary: Add/Remove movie from watchlist
      operationId: watchlist_movie_user_id_put
      parameters:
      - name: user_id
        in: path
        description: ID of the user
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          example: 42
      - name: id
        in: query
        description: ID of the movie
        required: true
        style: form
        explode: true
        schema:
          type: integer
          example: 42
      responses:
        "201":
          description: Title added/removed successfully
        "400":
          description: Bad request. Invalid user ID or title not provided
      x-openapi-router-controller: swagger_server.controllers.user_controller
  /watchlist/tvshow/{user_id}:
    put:
      tags:
      - User
      summary: Add/Remove TV show from watchlist
      operationId: watchlist_tvshow_user_id_put
      parameters:
      - name: user_id
        in: path
        description: ID of the user
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          example: 42
      - name: id
        in: query
        description: ID of TV show
        required: true
        style: form
        explode: true
        schema:
          type: integer
          example: 69
      responses:
        "201":
          description: Title added/removed successfully
        "400":
          description: Bad request. Invalid user ID or title not provided
      x-openapi-router-controller: swagger_server.controllers.user_controller
  /watchlist/movie/remove/{user_id}:
    put:
      tags:
      - User
      summary: Remove Movie from user's watchlist
      operationId: watchlist_movie_remove_user_id_put
      parameters:
      - name: user_id
        in: path
        description: ID of the user
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          example: 42
      - name: id
        in: query
        description: id of the Movie to be removed from user's watchlist
        required: true
        style: form
        explode: true
        schema:
          type: integer
          example: 42
      responses:
        "200":
          description: Movie removed successfully from user's wishlist
        "400":
          description: Bad request. Invalid user ID or missing required fields
      x-openapi-router-controller: swagger_server.controllers.user_controller
  /watchlist/tvshow/remove/{user_id}:
    put:
      tags:
      - User
      summary: Remove TV show from user's watchlist
      operationId: watchlist_tvshow_remove_user_id_put
      parameters:
      - name: user_id
        in: path
        description: ID of the user
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          example: 42
      - name: id
        in: query
        description: id of the TV show to be removed from user's watchlist
        required: true
        style: form
        explode: true
        schema:
          type: integer
          example: 69
      responses:
        "200":
          description: TV show removed successfully from user's wishlist
        "400":
          description: Bad request. Invalid user ID or missing required fields
      x-openapi-router-controller: swagger_server.controllers.user_controller
  /profile/{user_id}/username:
    put:
      tags:
      - User
      summary: Update username
      operationId: profile_user_id_username_put
      parameters:
      - name: user_id
        in: path
        description: ID of the user
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          example: 42
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/user_id_username_body'
      responses:
        "200":
          description: Username updated successfully
        "400":
          description: Bad request. Invalid user ID or missing required fields
        "404":
          description: User not found
      x-openapi-router-controller: swagger_server.controllers.user_controller
  /profile/{user_id}/pfp:
    put:
      tags:
      - User
      summary: Update profile picture
      operationId: profile_user_id_pfp_put
      parameters:
      - name: user_id
        in: path
        description: ID of the user
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          example: 42
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/user_id_pfp_body'
      responses:
        "200":
          description: Profile Picture updated successfully
        "400":
          description: Bad request. Invalid user ID or missing required fields
        "404":
          description: User not found
      x-openapi-router-controller: swagger_server.controllers.user_controller
  /profile/{user_id}/password:
    put:
      tags:
      - User
      summary: Update Password
      operationId: profile_user_id_password_put
      parameters:
      - name: user_id
        in: path
        description: ID of the user
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          example: 42
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/user_id_password_body'
      responses:
        "200":
          description: Password updated successfully
        "400":
          description: Bad request. Invalid user ID or missing required fields
        "404":
          description: User not found
      x-openapi-router-controller: swagger_server.controllers.user_controller
  /profile/{user_id}/interests:
    put:
      tags:
      - User
      summary: Update interests
      operationId: profile_user_id_interests_put
      parameters:
      - name: user_id
        in: path
        description: ID of the user
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          example: 42
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/user_id_interests_body'
      responses:
        "200":
          description: Interests updated successfully
        "400":
          description: Bad request. Invalid user ID or missing required fields
        "404":
          description: User not found
      x-openapi-router-controller: swagger_server.controllers.user_controller
  /profile/{user_id}/bio:
    put:
      tags:
      - User
      summary: Update user bio
      operationId: profile_user_id_bio_put
      parameters:
      - name: user_id
        in: path
        description: ID of the user
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          example: 42
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/user_id_bio_body'
      responses:
        "200":
          description: Bio updated successfully
        "400":
          description: Bad request. Invalid user ID or missing required fields
        "404":
          description: User not found
      x-openapi-router-controller: swagger_server.controllers.user_controller
components:
  schemas:
    title:
      type: object
      properties:
        movie:
          $ref: '#/components/schemas/movie'
        tvshow:
          $ref: '#/components/schemas/tvshow'
      example:
        movie:
          duration: 2hrs 2mins
          cast:
          - Leonardo Di Caprio
          - Kate Winslet
          - Bradd Pitt
          plot: plot
          release_date: 2001-08-19T00:00:00.000+00:00
          director: Christopher Nolan
          genres:
          - Adventure
          - Comedy
          - Thriller
          rating: 9.36
          writer: Todd Phillips
          movie_id:
            id: 42
          title: Inception
          poster: https://fmd.com/poster.jpeg
        tvshow:
          duration: 55mins
          cast:
          - Leonardo Di Caprio
          - Kate Winslet
          - Bradd Pitt
          show_id:
            id: 42
          plot: plot
          release_date: 2001-08-19T00:00:00.000+00:00
          director: Christopher Nolan
          genres:
          - Adventure
          - Comedy
          - Thriller
          rating: 9.36
          season: 2
          writer: Todd Phillips
          title: Peaky Blinders
          poster: https://fmd.com/poster.jpeg
    movie_id:
      type: object
      properties:
        id:
          type: integer
          example: 42
      example:
        id: 42
    show_id:
      type: object
      properties:
        id:
          type: integer
          example: 42
      example:
        id: 42
    user:
      type: object
      properties:
        user_id:
          type: integer
          example: 42
        username:
          maxLength: 25
          type: string
          example: bobdylan
        password:
          maxLength: 25
          type: string
          example: godofnoobs_42
        bio:
          maxLength: 100
          type: string
          example: "i am bobdylan, i work with duvylfyluksq"
        pfp:
          type: string
          format: uri
          example: https://fmd.com/pfp.jpeg
        watchlist_movies:
          type: object
          additionalProperties:
            type: boolean
          example:
            "1": true
            "2": false
            "42": true
        watchlist_shows:
          type: object
          additionalProperties:
            type: boolean
          example:
            "1": true
            "2": false
            "42": true
        interests:
          maxItems: 3
          minItems: 3
          type: array
          example:
          - Adventure
          - Comedy
          - Thriller
          items:
            type: string
      example:
        password: godofnoobs_42
        user_id: 42
        bio: "i am bobdylan, i work with duvylfyluksq"
        watchlist_movies:
          "1": true
          "2": false
          "42": true
        pfp: https://fmd.com/pfp.jpeg
        interests:
        - Adventure
        - Comedy
        - Thriller
        username: bobdylan
        watchlist_shows:
          "1": true
          "2": false
          "42": true
    movie:
      type: object
      properties:
        movie_id:
          $ref: '#/components/schemas/movie_id'
        title:
          type: string
          example: Inception
        plot:
          maxLength: 500
          type: string
        cast:
          maxItems: 3
          type: array
          example:
          - Leonardo Di Caprio
          - Kate Winslet
          - Bradd Pitt
          items:
            type: string
        director:
          type: string
          example: Christopher Nolan
        writer:
          type: string
          example: Todd Phillips
        rating:
          type: number
          format: float
          example: 9.36
        genres:
          maxItems: 3
          minItems: 3
          type: array
          example:
          - Adventure
          - Comedy
          - Thriller
          items:
            type: string
        release_date:
          type: string
          format: date
          example: 2001-08-19
        duration:
          pattern: ^\d+hrs \d+mins$
          type: string
          example: 2hrs 2mins
        poster:
          type: string
          example: https://fmd.com/poster.jpeg
      example:
        duration: 2hrs 2mins
        cast:
        - Leonardo Di Caprio
        - Kate Winslet
        - Bradd Pitt
        plot: plot
        release_date: 2001-08-19T00:00:00.000+00:00
        director: Christopher Nolan
        genres:
        - Adventure
        - Comedy
        - Thriller
        rating: 9.36
        writer: Todd Phillips
        movie_id:
          id: 42
        title: Inception
        poster: https://fmd.com/poster.jpeg
    tvshow:
      type: object
      properties:
        show_id:
          $ref: '#/components/schemas/show_id'
        title:
          type: string
          example: Peaky Blinders
        plot:
          maxLength: 500
          type: string
        season:
          type: integer
          example: 2
        cast:
          maxItems: 3
          type: array
          example:
          - Leonardo Di Caprio
          - Kate Winslet
          - Bradd Pitt
          items:
            type: string
        director:
          type: string
          example: Christopher Nolan
        writer:
          type: string
          example: Todd Phillips
        rating:
          type: number
          format: float
          example: 9.36
        genres:
          maxItems: 3
          minItems: 3
          type: array
          example:
          - Adventure
          - Comedy
          - Thriller
          items:
            type: string
        release_date:
          type: string
          format: date
          example: 2001-08-19
        duration:
          pattern: ^\d+mins$
          type: string
          example: 55mins
        poster:
          type: string
          example: https://fmd.com/poster.jpeg
      example:
        duration: 55mins
        cast:
        - Leonardo Di Caprio
        - Kate Winslet
        - Bradd Pitt
        show_id:
          id: 42
        plot: plot
        release_date: 2001-08-19T00:00:00.000+00:00
        director: Christopher Nolan
        genres:
        - Adventure
        - Comedy
        - Thriller
        rating: 9.36
        season: 2
        writer: Todd Phillips
        title: Peaky Blinders
        poster: https://fmd.com/poster.jpeg
...
    user_signin_body:
      required:
      - password
      - username
      type: object
      properties:
        username:
          maxLength: 25
          type: string
          example: bobdylan
        password:
          maxLength: 25
          type: string
          example: godofnoobs_42
    user_signup_body:
      required:
      - bio
      - confirm_password
      - interests
      - password
      - pfp
      - username
      type: object
      properties:
        username:
          maxLength: 25
          type: string
          example: bobdylan
        password:
          maxLength: 25
          type: string
          example: godofnoobs_42
        confirm_password:
          maxLength: 25
          type: string
          example: godofnoobs_42
        pfp:
          type: string
          example: https:example.com/pic.jpeg
        bio:
          type: string
          example: hi im bobdylan
        interests:
          maxItems: 3
          minItems: 3
          type: array
          items:
            type: string
    user_id_username_body:
      required:
      - username
      type: object
      properties:
        username:
          maxLength: 25
          type: string
    user_id_pfp_body:
      required:
      - pfp
      type: object
      properties:
        pfp:
          type: string
          example: https://www.example.com/1.jpeg
    user_id_password_body:
      required:
      - password
      type: object
      properties:
        password:
          maxLength: 25
          type: string
    user_id_interests_body:
      required:
      - interests
      type: object
      properties:
        interests:
          type: array
          items:
            maxItems: 3
            minItems: 3
            type: string
            example: "[\"Adventure\",\"Murder\",\"Gore\"]"
    user_id_bio_body:
      required:
      - bio
      type: object
      properties:
        bio:
          maxLength: 200
          type: string
...